@@ -14,24 +14,25 @@ namespace Web.App
     public class SpaSsrController : Controller
     {
         private readonly ILogger _logger;
-        private readonly IHostingEnvironment _env;
+        private readonly IWebHostEnvironment _env;
         private readonly IHttpClientFactory _httpClientFactory;
         private readonly IOptions<HypernovaSettings> _options;
-        private readonly HypernovaSettings _settings;
         private readonly IDistributedCache _cache;
-		private readonly IConfiguration _configuration;
 
 		private SpaSsr _spaSsr = null;
 
-        public SpaSsrController(ILogger<SpaSsrController> logger, IHttpClientFactory httpClientFactory, IHostingEnvironment env, IOptions<HypernovaSettings> options, IDistributedCache cache, IConfiguration configuration)
+        public SpaSsrController(ILogger<SpaSsrController> logger, IHttpClientFactory httpClientFactory, IWebHostEnvironment env, IOptions<HypernovaSettings> options, IDistributedCache cache, IConfiguration configuration)
         {
+            if (configuration is null)
+            {
+                throw new ArgumentNullException(nameof(configuration));
+            }
+
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
             _env = env ?? throw new ArgumentNullException(nameof(env));
             _httpClientFactory = httpClientFactory ?? throw new ArgumentNullException(nameof(httpClientFactory));
-            _options = options;
-            _settings = options.Value;
-            _cache = cache;
-			_configuration = configuration;
+            _options = options ?? throw new ArgumentNullException(nameof(options));
+            _cache = cache ?? throw new ArgumentNullException(nameof(cache));
 		}
 
         public async Task<ActionResult> Index()
@@ -47,11 +48,11 @@ namespace Web.App
 				}
 
 				var siteUrl = $"{protocol}://{host}";
-				_spaSsr = new SpaSsr(_logger, _env, _httpClientFactory, _options, _cache, siteUrl, _configuration, HttpContext);
+				_spaSsr = new SpaSsr(_logger, _env, _httpClientFactory, _options, _cache, siteUrl);
 
 			}
 
-            var cacheKey = HttpContext.Request.Path.ToString().ToLowerInvariant();
+            var cacheKey = HttpContext.Request.Path.ToString().ToUpperInvariant();
             var relativeUrl = $"{HttpContext.Request.Path}{HttpContext.Request.QueryString}";
 
             var renderResult = await _spaSsr.RenderSpaServerSide(cacheKey, relativeUrl, TimeSpan.FromDays(1.0));
